requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
Pillow
python-dotenv
